不记得具体改过什么了，由于要作较大的修改，在此记录一下目前的版本。

+++ source/stock/tcstockdata.h
@@ -10,10 +10,10 @@
 #define DAYS_OF_YEAR 366
 
 typedef struct {
-	qreal OpenPrice;
-	qreal ClosePrice;
-	qreal MaxPrice;
-	qreal MinPrice;
+	long OpenPrice;
+	long ClosePrice;
+	long MaxPrice;
+	long MinPrice;
 	long TotalPrice;
 	long Quantity;
 	long _Reserved1;	//not used in current version
